+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## 1.16.0
-### 03-Nov-2025
+### 05-Dec-2025
 - Added proof-of-concept validations for swiftDialog `2.5.1`'s "blurscreen" control
 - Removed vendor-specific Local Validations (in favor of Remote Validations)
 - Updated Configuration `policyJSON` to better match internal usage
@@ -10,8 +10,9 @@
 - Simplify Client-side Logging (thanks, @DevliegereM!)
 - Added proof-of-concept validations for swiftDialog `2.5.6`'s "hide or show" dialog window
 - Updated Dynamic Download Estimates for macOS 26 (and beyond)
-- Updated for swiftDialog 3.0.0
+- Updated for swiftDialog `3.0.0`
 - Updated `checkNetworkQualityCatchAllConfiguration` for macOS 26 (thanks for the heads-up, @Harald Brouwers!)
+- Added new salutation banner greeting, based on time o’ day (Pull Request #171; thanks, @ScottEKendall!)
 
 ## 1.15.1
 ### 06-Feb-2025